Project Euler, Problem 13
Find a number that has a lot of divisors. Had to do it piecemeal because of how slow is the algorithm. The algorithm can be improved by reducing the amount of numbers that have to do the MOD operation. The reduction can be achieved by checking if the number is divisible by a forbidden prime number where the forbidden prime number is a prime number with which the triangle number cannot be divided.